Dont clear user-mgnt on QA

+++ infrastructure/clear-all-data.sh
@@ -11,7 +11,7 @@
 set -e
 
 print_usage_and_exit () {
-    echo 'Usage: ./clear-all-data.sh REPLICAS'
+    echo 'Usage: ./clear-all-data.sh REPLICAS ENV'
     exit 1
 }
 
@@ -20,7 +20,13 @@ if [ -z "$1" ] ; then
     print_usage_and_exit
 fi
 
+if [ -z "$2" ] ; then
+    echo 'Error: Argument ENV is required in position 2.'
+    print_usage_and_exit
+fi
+
 REPLICAS=$1
+ENV=$2
 
 if [ "$REPLICAS" = "0" ]; then
   HOST=mongo1
@@ -49,7 +55,7 @@ docker run --rm --network=$NETWORK mongo:3.6 mongo hearth-dev --host $HOST --eva
 
 docker run --rm --network=$NETWORK mongo:3.6 mongo openhim-dev --host $HOST --eval "db.dropDatabase()"
 
-docker run --rm --network=$NETWORK mongo:3.6 mongo user-mgnt --host $HOST --eval "db.dropDatabase()"
+if [[ "$ENV" != "qa" ]] ; then docker run --rm --network=$NETWORK mongo:3.6 mongo user-mgnt --host $HOST --eval "db.dropDatabase()" ; fi
 
 docker run --rm --network=$NETWORK mongo:3.6 mongo application-config --host $HOST --eval "db.dropDatabase()"
